requirements.txt
flask==3.0.0
azure-identity==1.15.0
openai==1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
azure-keyvault-secrets==4.7.0
sqlparse==0.4.4
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
pandas==2.1.4
numpy==1.25.2
# Security packages
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
flask-talisman==1.1.0
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2
redis==5.0.1
# Additional data sources
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
